Group section by page, rename components
